# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: agent-fox hub REST API
  version: "0.19.0"
  summary: Headless harness for spec-driven, multi-agent software development.
  description: |
    The `hub` binary is the API server for agent-fox. It owns user identity,
    OAuth, workspaces, git hosting, the merge queue, carry-patch workflows,
    secrets, variables, agent sessions, and audit storage.

    ## Base paths

    | Prefix | Contents |
    |--------|----------|
    | `/api/v1` | Everything under the API mount point: workspaces, patches, rebuilds, merges, secrets, variables, sessions, audit, and the apikit identity and administration endpoints |
    | `/git` | Built-in git smart HTTP server (`/git/<org>/<slug>.git`) |
    | *(root)* | `/healthz`, `/readyz`, `/version`, and `/metrics` |

    The `/api/v1` prefix comes from `server.mount_point` in `config.toml` and
    defaults to `/api/v1`. Everything apikit mounts -- `/user`, `/users`,
    `/orgs`, `/auth` -- goes on that same group, so those paths are
    `/api/v1/user`, `/api/v1/users`, and so on. `docs/api.md` and
    `docs/permissions.md` both write them without the prefix; see
    `docs/errata/openapi_description.md`.

    ## Authentication

    Every `/api/v1` endpoint requires a bearer credential, except the two
    OAuth routes (`GET /api/v1/auth/providers` and
    `POST /api/v1/auth/callback`), which are registered before the auth
    middleware is attached to the group. Three credential types exist:

...
      identity to own them).
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{slug}/patches:
    parameters:
      - $ref: "#/components/parameters/WorkspaceSlug"
    post:
      tags: [Patches]
      operationId: addPatches
      summary: Add one or more patches
      description: |
        Accepts either a single patch object or a JSON array of them. A batch
        is validated and inserted inside one transaction: if any entry fails,
        nothing is inserted.

        When `position` is omitted the patch is appended at `max(position) + 1`.
        When it is supplied, existing patches at that position and beyond are
        shifted down by one. Positions greater than `max + 1` are clamped to an
        append.

        Branch existence in the workspace repository is validated at
        registration time unless `skip_branch_check` is `true`.

        Ownership is not enforced.
      x-required-scopes: ["patches:write"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/AddPatchRequest"
                - type: array
                  items:
                    $ref: "#/components/schemas/AddPatchRequest"
            examples:
              single:
                summary: Single patch
                value:
                  branch_name: feature/auth-headers
                  position: 2
                  upstream_pr_url: https://github.com/upstream/repo/pull/42
                  description: Add authentication headers to API
              batch:
                summary: Batch insert
                value:
                  - { branch_name: feature/a, position: 1 }
                  - { branch_name: feature/b }
                  - { branch_name: feature/c, upstream_pr_url: "https://github.com/upstream/repo/pull/99" }
      responses:
        "200":
          description: >-
            `if_not_exists` was true and the branch is already registered; the
            existing patch is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patch"
        "201":
          description: The created patch, or the array of created patches for a batch request.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Patch"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Patch"
        "400":
          description: >-
            `branch_name` is absent or empty; `branch_name` equals the
            workspace `integration_branch`; the workspace is missing, inactive,
            or in `standard` mode; `position` is below 1; or the branch does not
            exist in the repository and `skip_branch_check` is false.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "409":
          description: >-
            The branch is already registered for this workspace (and
            `if_not_exists` is false), or the batch contains a duplicate
            `branch_name`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: [Patches]
      operationId: listPatches
      summary: List patches
      description: |
        Returns every non-deleted patch ordered by ascending position.
        Soft-deleted patches (status `deleted`) are excluded. Standard
        (non-carry-patch) workspaces return an empty array rather than an
        error. Ownership is not enforced.
      x-required-scopes: ["patches:read", "patches:write"]
      responses:
        "200":
          description: The patch list in position order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patch"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/workspaces/{slug}/patches/{id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceSlug"
      - $ref: "#/components/parameters/PatchId"
    get:
      tags: [Patches]
      operationId: getPatch
      summary: Get a patch by ID
      description: Retrieve a single patch record by its UUID.
      x-required-scopes: ["patches:read"]
      responses:
        "200":
          description: The patch record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patch"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "404":
          $ref: "#/components/responses/PatchNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Patches]
      operationId: updatePatch
      summary: Update a patch
      description: |
        Only the supplied fields change. `branch_name` is immutable and is
        silently ignored if present. Changing `position` reshuffles the other
        patches so positions stay a contiguous 1-based sequence.
      x-required-scopes: ["patches:write"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatchRequest"
      responses:
        "200":
          description: The updated patch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patch"
        "400":
          description: >-
            `status` is not one of the valid values, or `position` is below 1 or
            above the total patch count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "404":
          $ref: "#/components/responses/PatchNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Patches]
      operationId: removePatch
      summary: Remove a patch
      description: |
        Deletes the patch row and compacts the remaining positions to a
        contiguous 1-based sequence. Both steps run in a single transaction.
      x-required-scopes: ["patches:write"]
      responses:
        "204":
          description: Patch removed and positions compacted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "404":
          $ref: "#/components/responses/PatchNotFound"
        "500":
          description: The delete-and-compact transaction failed; no partial state was persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/workspaces/{slug}/patches/reorder:
    parameters:
      - $ref: "#/components/parameters/WorkspaceSlug"
    post:
      tags: [Patches]
      operationId: reorderPatches
      summary: Reorder the whole patch list
      description: |
        `patch_ids` must list exactly the workspace's patch IDs -- no missing,
        extra, or duplicated entries. New positions are assigned from the array
        index (first element becomes position 1). The whole reorder runs in one
        transaction.
      x-required-scopes: ["patches:write"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderPatchesRequest"
      responses:
        "200":
          description: All patches in their new order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patch"
        "400":
          description: >-
            `patch_ids` is empty while patches exist, omits an existing patch,
            contains duplicates, references a patch from another workspace, or
            the bulk position update failed. Every reorder error surfaces as
            `400`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"

  # ---------------------------------------------------------------------------
  # Rerere
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <credential>`, where the credential is an admin
        token (`af_admin_<64-hex>`), an API key (`af_<key_id>_<secret>`), or a
        personal access token (`af_pat_<token_id>_<secret>`).
    gitBasicAuth:
      type: http
      scheme: basic
      description: |
        HTTP Basic authentication for the git smart HTTP endpoints. The
        username is ignored; the password must be a valid hub credential.
        `afc credential-helper get` returns the stored API key with the
        username `x-token-auth`.
  parameters:
    WorkspaceSlug:
      name: slug
      in: path
      required: true
      description: The workspace slug.
      schema:
        $ref: "#/components/schemas/Slug"
...
  responses:
    Unauthorized:
      description: No credential was supplied, or the credential is invalid, revoked, or expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: 401
              message: authentication required
    MissingScope:
      description: The PAT lacks a permission scope required by this endpoint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerError:
      description: An unexpected server-side failure, typically a database error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    WorkspaceNotFound:
      description: No workspace with this slug exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      description: The apikit error envelope used by every error response.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
              description: The HTTP status code, repeated in the body.
            message:
              type: string
              description: A human-readable description of the failure.
            error_type:
              type: string
              description: >-
                Machine-readable classification. Present only on the handlers
                that set it.
              enum:
                - duplicate_merge
                - workspace_mode_mismatch
                - no_active_patches
                - concurrent_rebuild
                - workspace_mismatch
                - workspace_access_denied
                - workspace_archived
                - unknown_schema_version
                - postmortem_not_found
      examples:
        - error:
            code: 400
            message: slug is required
        - error:
            code: 409
            message: a rebuild job is already queued or running for this workspace
            error_type: concurrent_rebuild
...
    Patch:
      type: object
      description: A patch branch carried on top of the upstream base.
      required:
        - id
        - workspace_slug
        - branch_name
        - position
        - status
        - upstream_pr_url
        - description
        - deleted_at
        - added_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workspace_slug:
          type: string
        branch_name:
          type: string
          description: The git branch backing this patch. Immutable after creation.
        position:
          type: integer
          minimum: 1
          description: 1-based application order; lower positions are applied first.
        status:
          $ref: "#/components/schemas/PatchStatus"
        conflict_files:
          type: array
          description: >-
            Conflicting file paths. Omitted entirely when the patch has no
            recorded conflicts.
          items:
            type: string
        upstream_pr_url:
          type: [string, "null"]
        description:
          type: [string, "null"]
        deleted_at:
          type: [string, "null"]
          format: date-time
          description: Soft-deletion timestamp; null for live patches.
        added_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        origin_sync_state:
          type: [string, "null"]
          enum: [in_sync, diverged, missing_on_origin, null]
          description: Fork-sync state. Null when unset or in hub mode.
        origin_sha:
          type: [string, "null"]
          description: Commit SHA on the fork remote at last sync. Null when unset or in hub mode.
        origin_synced_at:
          type: [string, "null"]
          format: date-time
          description: Timestamp of last fork-sync for this patch. Null when unset or in hub mode.
    AddPatchRequest:
      type: object
      required: [branch_name]
      properties:
        branch_name:
          type: string
          description: >-
            Must be non-empty, must differ from the workspace
            `integration_branch`, and must not already be registered.
        position:
          type: integer
          minimum: 1
          description: >-
            1-based insertion position. Values above `max(position) + 1` are
            clamped to an append. Omitted means append.
        upstream_pr_url:
          type: string
        description:
          type: string
        skip_branch_check:
          type: boolean
          default: false
          description: Skip the check that the branch exists in the repository.
        if_not_exists:
          type: boolean
          default: false
          description: >-
            When the branch is already registered, return `200` with the
            existing record instead of `409`.
...